Set up marker clustering
This commit creates a map that shows markers on every train track junction in White River Junction, VT.

When the user is zoomed in closely, the markers appear individually. When the user is zoomed out, markers that are in close proximity to each other are grouped together under a single marker by Google's MarkerClusterer library.